+++ src/core/thread/mle_router_ftd.hpp
@@ -419,7 +419,7 @@ public:
      * @param[out]  aChildInfo   The child information.
      *
      */
-    otError GetChildInfoByIndex(uint8_t aChildIndex, otChildInfo &aChildInfo);
+    otError GetChildInfoByIndex(uint16_t aChildIndex, otChildInfo &aChildInfo);
 
     /**
      * This methods gets the next IPv6 address (using an iterator) for a given child.
@@ -434,7 +434,7 @@ public:
      * @retval OT_ERROR_INVALID_ARGS  Child at @p aChildIndex is not valid.
      *
      */
-    otError GetChildNextIp6Address(uint8_t aChildIndex, Child::Ip6AddressIterator &aIterator, Ip6::Address &aAddress);
+    otError GetChildNextIp6Address(uint16_t aChildIndex, Child::Ip6AddressIterator &aIterator, Ip6::Address &aAddress);
 
     /**
      * This method indicates whether or not the RLOC16 is an MTD child of this device.
